backend/requirements.txt
openai
chromadb
PyMuPDF
tqdm
python-dotenv
langchain